Move as_ids to (new) sub-package scion
